requirements/base.txt
django
djangorestframework
psycopg2
django-environ
djangorestframework-simplejwt